requirements.txt
fastapi
sqlalchemy
passlib[bcrypt]
python-jose
uvicorn
python-dotenv
psycopg2-binary